+++ index.js
@@ -19,13 +19,16 @@ exports.boletoPayment = require('./sdk/boletoPayment');
 exports.utilityPayment = require('./sdk/utilityPayment');
 exports.paymentRequest = require('./sdk/paymentRequest');
 exports.webhook = require('./sdk/webhook');
+exports.workspace = require('./sdk/workspace');
 exports.event = require('./sdk/event');
 exports.key = require('./sdk/key.js');
 exports.error = require('./sdk/error.js');
+exports.organization = require('./sdk/user/organization.js')
 
 
 // Classes
 exports.Project = require('./sdk/user').Project;
+exports.Organization = require('./sdk/user').Organization;
 exports.Transaction = exports.transaction.Transaction;
 exports.Balance = exports.balance.Balance;
 exports.Boleto = exports.boleto.Boleto;
@@ -40,4 +43,5 @@ exports.UtilityPayment = exports.utilityPayment.UtilityPayment;
 exports.PaymentRequest = exports.paymentRequest.PaymentRequest;
 exports.Transfer = exports.transfer.Transfer;
 exports.Webhook = exports.webhook.Webhook;
+exports.Workspace = exports.workspace.Workspace;
 exports.Event = exports.event.Event;