+++ sdk/balance/balance.js
@@ -40,7 +40,8 @@ exports.get = async function ({user} = {}) {
      *
      * Parameters (optional):
      * @param user [Project object]: Project object. Not necessary if starkbank.user was set before function call
-     *
+     * @param user [Organization/Project object]: Organization or Project object. Not necessary if starkbank.user was set before function call
+     * 
      * Return:
      * @returns Balance object with updated attributes
      *