+++ sdk/workspace/index.js
@@ -1,0 +1,6 @@
+const workspace = require('./workspace');
+
+exports.create = workspace.create;
+exports.get = workspace.get;
+exports.query = workspace.query;
+exports.Workspace = workspace.Workspace;