+++ tests/testBrcodePreview.js
@@ -6,7 +6,7 @@ starkbank.user = require('./utils/user').exampleProject;
 describe('TestBrcodePreview', function(){
     this.timeout(10000);
     it('test_success', async () => {
-        const brcode = "00020126580014br.gov.bcb.pix0136a629532e-7693-4846-852d-1bbff817b5a8520400005303986540510.005802BR5908T'Challa6009Sao Paulo62090505123456304B14A";
+        const brcode = "00020126580014br.gov.bcb.pix013635719950-ac93-4bab-8ad6-56d7fb63afd252040000530398654040.005802BR5915Stark Bank S.A.6009Sao Paulo62070503***6304AA26";
         const previews = await starkbank.brcodePreview.query({brcodes: [brcode]});
         for await (let preview of previews) {
             assert(typeof preview.status == 'string');